% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_spde.R
\name{bru_graph_rep}
\alias{bru_graph_rep}
\title{Creates a vector of replicates to be used with 'inlabru'}
\usage{
bru_graph_rep(repl, graph_spde)
}
\arguments{
\item{repl}{A vector of replicates. If set to \verb{__all}, a vector
for all replicates will be generated.}

\item{graph_spde}{Name of the field.}
}
\value{
A vector of replicates to be used with 'inlabru'.
}
\description{
Auxiliary function to create a vector of replicates to be used
with 'inlabru'.
}
