% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.prec.R
\name{build.C.beta1}
\alias{build.C.beta1}
\title{computes C matrix for beta=1}
\usage{
build.C.beta1(L, kappa, sigma)
}
\arguments{
\item{L}{edge length}

\item{kappa}{matern parameter}

\item{sigma}{matern parameter}

\item{nu}{matern parameter}
}
\description{
computes C matrix for beta=1
}
