% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.prec.R
\name{build.Sigma.beta1}
\alias{build.Sigma.beta1}
\title{computes matrices needed for graph for beta=1}
\usage{
build.Sigma.beta1(P, E, kappa, sigma, n)
}
\arguments{
\item{P}{vertices}

\item{E}{edges}

\item{kappa}{matern parameter}

\item{sigma}{matern parameter}

\item{n}{number of nodes to include per edge}
}
\value{
a list with elements:
    Sigma : Covariance matrix of 'free' processes and derivatives on the edges
    A : Matrix that implements the Kirchoff constraints
    rep.ind : the indices of the repeated nodes
    p.ind : the indices in Sigma corresponding to the process
    d.ind : the indices in Sigma corresponding to the derivative
    P : matrix with the locations of the nodes where Sigma is evaluated
}
\description{
computes matrices needed for graph for beta=1
}
