% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.fem.R
\name{compute.lengths}
\alias{compute.lengths}
\title{Compute edge lengths in graph}
\usage{
compute.lengths(P, E)
}
\arguments{
\item{P}{Vertex matrix}

\item{E}{Edge matrix}
}
\value{
Vector with lengths
}
\description{
Compute edge lengths in graph
}
