% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_length}
\alias{compute_length}
\title{Compute length}
\usage{
compute_length(coords)
}
\arguments{
\item{coords}{nx2 matrix Matrix of the points of the lines}
}
\description{
Computes the length of a piecewise-linear function whose coordinates are given in a matrix.
}
