% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.prec.R
\name{corrector.inverse.e}
\alias{corrector.inverse.e}
\title{inverse corrector elements
builds the elements of the inverse of the corrector matrix
r(t,s)  =r_0(t-s) - phi(t) A phi(s)
where r_0 is stationary matern}
\usage{
corrector.inverse.e(kappa, sigma, nu = 3/2, L = 1)
}
\arguments{
\item{kappa}{matern param}

\item{sigma}{matern param}

\item{nu}{shape param}

\item{L}{interval length}
}
\description{
inverse corrector elements
builds the elements of the inverse of the corrector matrix
r(t,s)  =r_0(t-s) - phi(t) A phi(s)
where r_0 is stationary matern
}
