% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.prec.R
\name{corrector}
\alias{corrector}
\title{The corrector matrix, A, such that neumann is:
r(t,s)  =r_0(t-s) - phi(t) A phi(s)
where r_0 is stationary matern}
\usage{
corrector(kappa, sigma, nu = 3/2, L = 1)
}
\arguments{
\item{kappa}{matern param}

\item{sigma}{matern param}

\item{nu}{shape param}

\item{L}{interval length}
}
\description{
The corrector matrix, A, such that neumann is:
r(t,s)  =r_0(t-s) - phi(t) A phi(s)
where r_0 is stationary matern
}
