% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expontial.covariance.R
\name{covariance.point.to.graph.exp}
\alias{covariance.point.to.graph.exp}
\title{Compute covariance of a point to the entire graph (discretized)}
\usage{
covariance.point.to.graph.exp(EP, theta, graph, n.p = 50)
}
\arguments{
\item{EP}{- (2 x 1) [1] -edge number, [2] -normalized location on the edge}

\item{theta}{- (3 x 1) (kappa,sigma)}

\item{graph}{- (graph)}

\item{n.p}{- (int) number of points to compute the covariance on on each edge}
}
\value{
C      - (n.p*numer of edges x 3) [,1] edge number [,2] length from lower edge [,3] covariance
}
\description{
Compute covariance of a point to the entire graph (discretized)
}
