% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance_alpha1.R
\name{covariance_alpha1}
\alias{covariance_alpha1}
\title{Compute covariance of a point to the entire graph (discretized) for
alpha=1 model}
\usage{
covariance_alpha1(P, kappa, tau, range, sigma, graph, n.p = 50, scale = FALSE)
}
\arguments{
\item{P}{(2 x 1) point with edge number and normalized location on the edge}

\item{kappa}{parameter kappa from the SPDE}

\item{tau}{parameter tau from the SPDE}

\item{range}{range parameter}

\item{sigma}{std. deviation parameter}

\item{graph}{metric_graph object}

\item{n.p}{number of points to compute the covariance on each edge}

\item{scale}{scale the covariance by \code{2*kappa} so that sigma corresponds to
the marginal standard deviation (default FALSE)}
}
\value{
C (n.p*number of edges x 3) \verb{[,1]} edge number \verb{[,2]} distance from
lower edge \verb{[,3]} covariance
}
\description{
Compute covariance of a point to the entire graph (discretized) for
alpha=1 model
}
