% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance_alpha2.R
\name{covariance_alpha2_mesh}
\alias{covariance_alpha2_mesh}
\title{Compute covariance of a point to the mesh locations in a graph for
alpha=2 model}
\usage{
covariance_alpha2_mesh(P, kappa, tau, range, sigma, graph)
}
\arguments{
\item{P}{(2 x 1) point with edge number and normalized location on the edge}

\item{kappa}{parameter kappa from the SPDE}

\item{tau}{parameter tau from the SPDE}

\item{range}{range parameter}

\item{sigma}{std. deviation parameter}

\item{graph}{metric_graph object}
}
\value{
a vector with covariance values
}
\description{
Compute covariance of a point to the mesh locations in a graph for
alpha=2 model
}
