% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{drop_na.metric_graph_data}
\alias{drop_na.metric_graph_data}
\alias{drop_na}
\title{A version of \code{tidyr::drop_na()} function for datasets on metric graphs}
\usage{
\method{drop_na}{metric_graph_data}(data, ...)
}
\arguments{
\item{data}{The data list or \code{tidyr::tibble} obtained from a metric graph object.}

\item{...}{Additional parameters to be passed to \code{tidyr::drop_na()}.}
}
\value{
A \code{tidyr::tibble} with the resulting selected columns.
}
\description{
Applies \code{tidyr::drop_na()} function for datasets obtained from a metric graph object.
}
