% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{exp_covariance}
\alias{exp_covariance}
\title{Exponential covariance function}
\usage{
exp_covariance(h, theta)
}
\arguments{
\item{h}{Distances to evaluate the covariance function at.}

\item{theta}{A vector \code{c(sigma, kappa)}, where \code{sigma} is the standard
deviation and \code{kappa} is a range-like parameter.}
}
\value{
A vector with the values of the covariance function.
}
\description{
Evaluates the exponential covariance function
\deqn{C(h) = \sigma^2 \exp\{-kappa h\}}
}
