% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_spde.R
\name{gg_df.metric_graph_spde_result}
\alias{gg_df.metric_graph_spde_result}
\alias{gg_df}
\title{Data frame for metric_graph_spde_result objects to be used in 'ggplot2'}
\usage{
\method{gg_df}{metric_graph_spde_result}(
  result,
  parameter = result$params,
  transform = TRUE,
  restrict_x_axis = parameter,
  restrict_quantiles = list(sigma = c(0, 1), range = c(0, 1), kappa = c(0, 1), sigma =
    c(0, 1)),
  ...
)
}
\arguments{
\item{result}{A metric_graph_spde_result object.}

\item{parameter}{Vector. Which parameters to get the posterior density in the
data.frame? The options are \code{sigma}, \code{range} or \code{kappa}.}

\item{transform}{Should the posterior density be given in the original scale?}

\item{restrict_x_axis}{Variables to restrict the range of x axis based on quantiles.}

\item{restrict_quantiles}{List of quantiles to restrict x axis.}

\item{...}{Not being used.}
}
\value{
A \code{data.frame} containing the posterior densities.
}
\description{
Returns a 'ggplot2'-friendly data-frame with the marginal posterior densities.
}
