% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_spde.R
\name{gpgraph_spde}
\alias{gpgraph_spde}
\title{INLA implementation of Whittle-Matérn fields for metric graphs.}
\usage{
gpgraph_spde(
  graph_object,
  alpha = 1,
  stationary_endpoints = "all",
  parameterization = c("matern", "spde"),
  start_range = NULL,
  prior_range = NULL,
  start_kappa = NULL,
  start_sigma = NULL,
  prior_kappa = NULL,
  prior_sigma = NULL,
  debug = FALSE
)
}
\arguments{
\item{graph_object}{A `metric_graph` object.}

\item{alpha}{The order of the SPDE.}

\item{stationary_endpoints}{Which vertices of degree 1 should contain stationary boundary conditions?}

\item{parameterization}{Which parameterization to be used? The options are 'matern' (sigma and range) and 'spde' (sigma and kappa).}

\item{start_range}{Starting value for range parameter.}

\item{start_kappa}{Starting value for kappa.}

\item{start_sigma}{Starting value for sigma.}

\item{debug}{Should debug be displayed?}

\item{prior.range}{a `list` containing the elements `meanlog` and
`sdlog`, that is, the mean and standard deviation on the log scale. Will not be used if prior.kappa is non-null.}

\item{prior.kappa}{a `list` containing the elements `meanlog` and
`sdlog`, that is, the mean and standard deviation on the log scale.}

\item{prior.sigma}{a `list` containing the elements `meanlog` and
`sdlog`, that is, the mean and standard deviation on the log scale.}
}
\value{
An inla object.
}
\description{
This function creates an inla object that can be used
in `INLA` or `inlabru` to fit Whittle-Matérn fields 
on metric graphs.
}
