% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.fem.R
\name{graph.mesh}
\alias{graph.mesh}
\title{compute the mesh with a given mesh width h}
\usage{
graph.mesh(P, E, h, n = NULL)
}
\arguments{
\item{P}{Vertex matrix of graph}

\item{E}{Edge matrix of graph}

\item{h}{desired mesh width}

\item{n}{desired number of vertices per edge in the graph}
}
\value{
list
}
\description{
compute the mesh with a given mesh width h
}
