% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_spde.R
\name{graph_bru_process_data}
\alias{graph_bru_process_data}
\title{Prepare data frames or data lists to be used with 'inlabru' in metric
graphs}
\usage{
graph_bru_process_data(
  data,
  edge_number = "edge_number",
  distance_on_edge = "distance_on_edge",
  loc = "loc"
)
}
\arguments{
\item{data}{A \code{data.frame} or a \code{list} containing the covariates, the edge
number and the distance on edge for the locations to obtain the prediction.}

\item{edge_number}{Name of the variable that contains the edge number, the
default is \code{edge_number}.}

\item{distance_on_edge}{Name of the variable that contains the distance on
edge, the default is \code{distance_on_edge}.}

\item{loc}{character. Name of the locations to be used in 'inlabru' component.}
}
\value{
A list containing the processed data to be used in a user-friendly
manner by 'inlabru'.
}
\description{
Prepare data frames or data lists to be used with 'inlabru' in metric
graphs
}
