% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_lgcp.R
\name{graph_lgcp}
\alias{graph_lgcp}
\title{Simulation of log-Gaussian Cox process on metric graph}
\usage{
graph_lgcp(intercept = 0, sigma, range, alpha, graph)
}
\arguments{
\item{intercept}{mean value of the Gaussian process}

\item{sigma}{parameter for marginal standard deviations}

\item{range}{parameter for practical correlation range}

\item{alpha}{smoothness parameter (1 or 2)}

\item{graph}{metric_graph object}
}
\value{
List with Gaussian process sample and simulated points
}
\description{
Simulation of log-Gaussian Cox process on metric graph
}
