% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_lme.R
\name{graph_lme}
\alias{graph_lme}
\title{Metric graph linear mixed effects models}
\usage{
graph_lme(
  formula,
  graph,
  model = list(type = "WhittleMatern", alpha = 1),
  repl = NULL,
  optim_method = "L-BFGS-B",
  parallel = FALSE,
  parallel_controls = list(n_cores = parallel::detectCores() - 1),
  optim_controls = list()
)
}
\arguments{
\item{formula}{Formula object describing the relation between the response variables and the fixed effects.}

\item{graph}{A \code{metric_graph} object.}

\item{model}{The random effects model that will be used. A list containing the elements \code{type}, which can be
\code{Whittle-Matern}, \code{graph-Laplacian} or \code{isoCov}. For \code{Whittle-Matern} and \code{graph-Laplacian} models, the list
must also contain a parameter \code{alpha} (which is 1 by default). For \code{isoCov} models, the list must
contain a parameter \code{cov_function}, containing the covariance function (which is \code{exp_covariance}, the
exponential covariance, by default). Finally, for \code{Whittle-Matern} models, there is an additional parameter
\code{version}, which can be either 1 or 2, to tell which version of the likelihood should be used (version is 1 by default).}

\item{repl}{A vector containing the replicates.}

\item{optim_method}{The method to be used with \code{optim} function.}

\item{parallel}{Logical. Should optimParallel be used instead of optim?}

\item{parallel_controls}{A list containing the parallel controls.}

\item{optim_controls}{Additional controls to be passed to \code{optim} or \code{optimParallel}.}
}
\value{
A list containing the fitted model.
}
\description{
Fitting linear mixed effects model in metric graphs. The random effects can be
Whittle-Matern fields on metric graphs, Whittle-Matern fields based on the
graph Laplacian as well as fields with isotropic covariance structure.
}
