% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance_alpha2.R
\name{graph_posterior_mean_matern2}
\alias{graph_posterior_mean_matern2}
\title{Generates samples of the entire graph}
\usage{
graph_posterior_mean_matern2(graph, theta, sample = F)
}
\arguments{
\item{graph}{- graphical object}

\item{theta}{- (sigma_e, sigma, kappa)}
}
\description{
Generates samples of the entire graph
}
