% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_spde.R
\name{graph_stack}
\alias{graph_stack}
\title{Creates an inla.stack for metric graph objects}
\usage{
graph_stack(stack_obj, name)
}
\arguments{
\item{stack_obj}{A density data frame}

\item{name}{Name of the field}
}
\value{
A data object to be passed to inla
}
\description{
Auxiliar function to create inla.stacks for metric graph objects
}
