% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_spde.R
\name{inlabru_predict}
\alias{inlabru_predict}
\title{Kriging with inlabru}
\usage{
inlabru_predict(
  bru_model,
  bru_fit,
  cmp,
  XY = NULL,
  PtE = NULL,
  data_pred = NULL
)
}
\arguments{
\item{bru_model}{An \code{inla_metric_graph_spde} object built with the \code{graph_spde()} function.}

\item{bru_fit}{A fitted model using \code{inlabru} or \code{inla}.}

\item{cmp}{An inlabru component.}

\item{XY}{Euclidean coordinates of the prediction locations.}

\item{PtE}{Relative positions on the edge to obtain predictions.}

\item{data_pred}{A data.frame containing the prediction locations along with the response variables (with NA if they are missing).}
}
\value{
A list with predictions.
}
\description{
Auxiliar function to obtain predictions of the field
using inlabru.
}
