% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expontial.covariance.R
\name{likelihood.exp.graph.covariance}
\alias{likelihood.exp.graph.covariance}
\title{Likelihood evaluation not using sparsity}
\usage{
likelihood.exp.graph.covariance(theta, graph.obj)
}
\arguments{
\item{theta}{- (sigma_e, sigma, kappa)}

\item{graph.obj}{- graph object}
}
\value{
The log-likelihood
}
\description{
Likelihood evaluation not using sparsity
}
