% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expontial.covariance.R
\name{likelihood.exp.graph.v2}
\alias{likelihood.exp.graph.v2}
\title{Log-likelihood calculation for alpha=1 without integration}
\usage{
likelihood.exp.graph.v2(theta, graph.obj)
}
\arguments{
\item{theta}{- (sigma_e, sigma, kappa)}

\item{graph.obj}{- graph object}
}
\value{
The log-likelihood
}
\description{
Log-likelihood calculation for alpha=1 without integration
}
