% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_inference.R
\name{likelihood.graph.covariance}
\alias{likelihood.graph.covariance}
\title{Likelihood evaluation not using sparsity}
\usage{
likelihood.graph.covariance(theta, graph, model = "alpha1")
}
\arguments{
\item{theta}{- (sigma_e, sigma, kappa)}

\item{model}{Model for Gaussian process, supported options are alpha1 (SPDE with alpha=1),
alpha2 (SPDE with alpha=2), GL (graph Laplacian model with alpha=1), GL2 (graph Laplacian
model with alpha=2) and isoExp (model with isotropic exponential covariance)}

\item{graph.obj}{- graph object}
}
\value{
The log-likelihood
}
\description{
Likelihood evaluation not using sparsity
}
