% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matern2.covariance.R
\name{likelihood.matern2.graph}
\alias{likelihood.matern2.graph}
\title{Computes the log likelihood function fo theta for the graph object}
\usage{
likelihood.matern2.graph(theta, graph.obj)
}
\arguments{
\item{theta}{- (sigma_e, sigma, kappa)}

\item{graph.obj}{- graphical object}
}
\description{
Computes the log likelihood function fo theta for the graph object
}
