% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_likelihoods.R
\name{likelihood_alpha1}
\alias{likelihood_alpha1}
\title{Log-likelihood calculation for alpha=1 model}
\usage{
likelihood_alpha1(theta, graph)
}
\arguments{
\item{theta}{(sigma_e, sigma, kappa)}

\item{graph}{metric_graph object}
}
\description{
Log-likelihood calculation for alpha=1 model
}
