% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_likelihoods.R
\name{likelihood_alpha1_v2}
\alias{likelihood_alpha1_v2}
\title{Log-likelihood calculation for alpha=1 model}
\usage{
likelihood_alpha1_v2(theta, graph)
}
\arguments{
\item{theta}{(sigma_e, sigma, kappa)}

\item{graph}{metric_graph object}
}
\value{
The log-likelihood
}
\description{
Log-likelihood calculation for alpha=1 model
}
\details{
This function computes the likelihood without integrating out
the vertices.
}
