% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_likelihoods.R
\name{likelihood_alpha2}
\alias{likelihood_alpha2}
\title{Computes the log likelihood function fo theta for the graph object}
\usage{
likelihood_alpha2(theta, graph)
}
\arguments{
\item{theta}{- (sigma_e, sigma, kappa)}

\item{graph.obj}{- graphical object}
}
\description{
Computes the log likelihood function fo theta for the graph object
}
