% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_likelihoods.R
\name{likelihood_graph_covariance}
\alias{likelihood_graph_covariance}
\title{Likelihood evaluation not using sparsity}
\usage{
likelihood_graph_covariance(theta, graph, model = "alpha1")
}
\arguments{
\item{theta}{parameters (sigma_e, sigma, kappa)}

\item{graph}{metric_graph object}

\item{model}{Type of model: "alpha1" gives SPDE with alpha=1, "GL1" gives
the model based on the graph Laplacian with smoothness 1, "GL2" gives the
model based on the graph Laplacian with smoothness 2, and "isoExp" gives a
model with isotropic exponential covariance#' @return The log-likelihood}
}
\description{
Likelihood evaluation not using sparsity
}
