% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_likelihoods.R
\name{likelihood_graph_laplacian}
\alias{likelihood_graph_laplacian}
\title{Log-likelihood calculation for graph Laplacian model with alpha=1}
\usage{
likelihood_graph_laplacian(theta, graph, alpha)
}
\arguments{
\item{theta}{(sigma_e, sigma, kappa)}

\item{graph}{metric graph object}

\item{alpha}{integer values supported}
}
\value{
The log-likelihood
}
\description{
Log-likelihood calculation for graph Laplacian model with alpha=1
}
