% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psp_to_metricgraph.R
\name{linnet.to.graph}
\alias{linnet.to.graph}
\title{from
function to export a linnet object, from spatstat package, to metric graph object}
\usage{
linnet.to.graph(linnet.object)
}
\description{
from
function to export a linnet object, from spatstat package, to metric graph object
}
