% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/space_time.R
\name{make.Q.spacetime}
\alias{make.Q.spacetime}
\title{Space-time precision operator discretization}
\usage{
make.Q.spacetime(graph, t, kappa, rho, gamma, alpha, beta, sigma)
}
\arguments{
\item{graph}{metric_graph object}

\item{t}{vector of time points}

\item{kappa}{range parameter kappa}

\item{rho}{drift parameter}

\item{gamma}{temporal range parameter}

\item{alpha}{smoothness parameter (integer) for spatial operator}

\item{beta}{smoothness parameter (integer) for Q-Wiener process}

\item{sigma}{variance parameter}
}
\value{
Precision matrix
}
\description{
The precision matrix for all vertices for space-time field
}
