% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.prec.R
\name{matern_derivative}
\alias{matern_derivative}
\title{Derivatives of the Matern covariance}
\usage{
matern_derivative(h, kappa, nu, sigma, deriv = 1)
}
\arguments{
\item{h}{distances where the covariance should be evaluated}

\item{kappa}{range parameter}

\item{nu}{smoothness parameter}

\item{sigma}{standard deviation}

\item{deriv}{order of derivative}
}
\value{
covariance function
}
\description{
Derivatives of the Matern covariance
}
