% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.prec.R
\name{matern.neumann.c}
\alias{matern.neumann.c}
\title{computes the covariance of r(s,t) using the corrector matrix form}
\usage{
matern.neumann.c(s, t, kappa, sigma, nu = 3/2, L = 1, deriv = c(0, 0))
}
\arguments{
\item{s}{(n x 1) location}

\item{t}{(n x 1) location}

\item{kappa}{matern param}

\item{sigma}{matern param}

\item{nu}{shape param}

\item{L}{interval length}

\item{deriv}{derivative order s, t}
}
\description{
computes the covariance of r(s,t) using the corrector matrix form
}
