% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.prec.R
\name{matern.neumann.free}
\alias{matern.neumann.free}
\title{computes the covariance of free space (r'=0) neumann boundary}
\usage{
matern.neumann.free(s, t, kappa, sigma, nu = 3/2, L = 1, deriv = c(0, 0))
}
\arguments{
\item{s}{(n x 1) location}

\item{t}{(n x 1) location}
}
\description{
computes the covariance of free space (r'=0) neumann boundary
}
