% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.prec.R
\name{matern.neumann.free2}
\alias{matern.neumann.free2}
\title{computes the covariance of free space (r'=0) neumann boundary with arbitrary
covarians on X'}
\usage{
matern.neumann.free2(
  s,
  t,
  C,
  kappa,
  sigma = 1,
  nu = 3/2,
  L = 1,
  deriv = c(0, 0)
)
}
\arguments{
\item{s}{(n x 1) location}

\item{t}{(n x 1) location}

\item{C}{(2 x 2) covarians of X'}

\item{kappa}{matern param}

\item{sigma}{matern param}

\item{nu}{shape param}

\item{L}{interval length}

\item{deriv}{derivative order s, t}
}
\description{
computes the covariance of free space (r'=0) neumann boundary with arbitrary
covarians on X'
}
