% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{matern_neumann_free}
\alias{matern_neumann_free}
\title{computes the covariance of free space (r'=0) neumann boundary}
\usage{
matern_neumann_free(s, t, kappa, sigma, nu = 3/2, L = 1, deriv = c(0, 0))
}
\arguments{
\item{s}{(n x 1) location}

\item{t}{(n x 1) location}

\item{kappa}{a range-like parameter (from the SPDE)}

\item{sigma}{the standard deviation}

\item{nu}{the smoothness parameter}

\item{L}{interval length}

\item{deriv}{a vector containing the order of the derivatives}
}
\description{
computes the covariance of free space (r'=0) neumann boundary
}
