% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetricGraph.R
\docType{package}
\name{MetricGraph}
\alias{MetricGraph}
\title{Gaussian processes on metric graphs}
\description{
\code{MetricGraph} is used for specifying Gaussian random fields on metric graphs.
The main models are the Whittle-Matérn fields, which are specified through the
fractional elliptic SPDE
\deqn{(\kappa^2 - \Delta)^{\alpha/2} (\tau u(s)) = W,}
\eqn{\kappa,\tau>0} and \eqn{\alpha>1/2} are parameters and \eqn{W} is Gaussian white noise.
The package also implements models based on graph Laplacians and isotropic covariance functions,
and contains various utility functions for specifying graphs, computing likelihoods, performing
prediction, simulating processes, and visualizing results and graphs.
}
\details{
At the heart of the package is the \code{R6} class \verb{[metric_graph()]}. This is used for specifying
metric graphs, and contains various utility functions which are needed for specifying Gaussian
processes on such spaces.

Basic statistical operations such as likelihood evaluations (see
\verb{[likelihood.graph.covariance], [likelihood.graph]}) and
predictions (see ...) are also implemented...

For a more detailed introduction to the package, see the MetricGraph Vignettes.
}
