% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.prec.R
\name{phi}
\alias{phi}
\title{one dim corrector function}
\usage{
phi(t, kappa, sigma, nu = 3/2, L = 1, deriv = 0)
}
\arguments{
\item{t}{(n x 1) where to evaluate phi (in \link{0,L})}

\item{kappa}{matern param}

\item{sigma}{matern param}

\item{nu}{shape param}

\item{L}{interval length}

\item{deriv}{how many derivatives}
}
\value{
phi - (n x m)
}
\description{
one dim corrector function
}
