% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.prec.R
\name{phi.to.phid}
\alias{phi.to.phid}
\title{Transforms the boundary conditions to phi
phi'(t) = B phi(t)}
\usage{
phi.to.phid(kappa, sigma, nu = 3/2, L = 1)
}
\description{
Transforms the boundary conditions to phi
phi'(t) = B phi(t)
}
