% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.plot.R
\name{plot_graph}
\alias{plot_graph}
\title{Plots of graphs}
\usage{
plot_graph(
  P,
  E,
  f = NULL,
  nc = 100,
  rgl = FALSE,
  N.e = NULL,
  add = FALSE,
  p = NULL,
  col = NULL,
  size = 6,
  plot.vertices = TRUE
)
}
\arguments{
\item{P}{Vertex matrix of graph}

\item{E}{Edge matrix of graph}

\item{f}{function to plot}

\item{nc}{number of colors to use}

\item{rgl}{Use RGL plot?}

\item{N.e}{Number of points at which the function is define on each edge}

\item{add}{Add plot to existing plot p (for RGL)?}

\item{p}{Plot to add results to}

\item{col}{color to use}

\item{size}{size of elements to plot}

\item{plot.vertices}{plot the vertices of the graph?}
}
\value{
The plot object
}
\description{
Plots of graphs
}
