% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.plot.R
\name{plot_posterior_mean}
\alias{plot_posterior_mean}
\title{(depricated)}
\usage{
plot_posterior_mean(
  graph,
  posterior.mean,
  sample.line,
  theta,
  byVertex = T,
  X.intercept = 0,
  ...
)
}
\arguments{
\item{graph}{- graph.obj}

\item{posterior.mean}{- (function) takes theta, and graph computes posterior mean of the verteces}

\item{sample.line}{- (function) takes theta, and graph computes posterior mean}

\item{theta}{- params}

\item{byVertex}{- if byVertex V.i is by vertex ortherwise by edge}
}
\description{
(depricated)
}
