% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expontial.covariance.R
\name{plot_r_1}
\alias{plot_r_1}
\title{plot the exponential covariance for parameter set}
\usage{
plot_r_1(theta, t = NULL)
}
\arguments{
\item{theta}{- sigma_e, kappa, sigma}
}
\description{
plot the exponential covariance for parameter set
}
