% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spde_precision.R
\name{plot_r_2}
\alias{plot_r_2}
\title{plot the Matern alpha=2 covariance for parameter set}
\usage{
plot_r_2(theta, t = NULL)
}
\arguments{
\item{theta}{- sigma_e, kappa, sigma}
}
\description{
plot the Matern alpha=2 covariance for parameter set
}
