% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/space_time.R
\name{plot_spacetime_covariances}
\alias{plot_spacetime_covariances}
\title{Function for plotting space-time covariances}
\usage{
plot_spacetime_covariances(
  graph,
  Q,
  Qs = NULL,
  t.ind,
  s.ind,
  t.shift = 0,
  show.temporal = FALSE,
  t
)
}
\arguments{
\item{graph}{metric_graph object}

\item{Q}{space-time precision matrix}

\item{Qs}{Spatial precision matrix as reference}

\item{t.ind}{time indices to plot covariances for}

\item{s.ind}{space indices to plot covariances for}

\item{t.shift}{time shifts to plot, i.e., covariances C(u(s(s.ind),t),u(*,t+t.shift))
are shown.}

\item{show.temporal}{Plot the marginal temporal covariances?}

\item{t}{vector with the time points for which Q is computed}
}
\description{
Plots of marginal spatial and temporal covariances of spacetime model
}
