% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.plot.R
\name{plot_straight_curve}
\alias{plot_straight_curve}
\title{plot a straight curve curve}
\usage{
plot_straight_curve(data.to.plot, V, gg = NULL, ...)
}
\arguments{
\item{data.to.plot}{(m x 2) [,1] position on curve (in length)
[,2] value}

\item{V}{(2 x 2) the position of the two vertices in 2D}

\item{gg}{(ggplot pbj) add to gg obj if not null}
}
\description{
plot a straight curve curve
}
