% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.plot.R
\name{plot_X_to_graph}
\alias{plot_X_to_graph}
\title{plot generic data object X on the graph}
\usage{
plot_X_to_graph(X, graph, Lines = NULL, gg = NULL, ...)
}
\arguments{
\item{X}{(m x 3)  [, 1] edge number
[, 2] position on curve (in length)
[, 3] value}

\item{graph}{(graph.object)}

\item{gg}{(ggplot pbj) add to gg obj if not null}
}
\description{
plot generic data object X on the graph
}
