% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expontial.covariance.R
\name{poster.mean.exp}
\alias{poster.mean.exp}
\title{Computes the posterior expectation for each node in the graph}
\usage{
poster.mean.exp(theta, graph.obj, rem.edge = NULL)
}
\arguments{
\item{theta}{- (sigma_e, sigma, kappa)}

\item{graph.obj}{- graphical object}

\item{rem.edge}{- remove edge}
}
\description{
Computes the posterior expectation for each node in the graph
}
