% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matern2.covariance.R
\name{poster.mean.obs.matern2}
\alias{poster.mean.obs.matern2}
\title{Computes the posterior expectation for each observation in the graph}
\usage{
poster.mean.obs.matern2(theta, graph.obj, leave.edge.out = F)
}
\arguments{
\item{theta}{- (sigma_e, sigma, kappa)}

\item{graph.obj}{- graphical object}

\item{leave.edge.out}{- compute the expectation of the graph if the observatrions are not on the edge}
}
\description{
Computes the posterior expectation for each observation in the graph
}
