% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expontial.covariance.R
\name{posterior.leave.graph_laplacian}
\alias{posterior.leave.graph_laplacian}
\title{Prediction for the graph-Laplacian model}
\usage{
posterior.leave.graph_laplacian(theta, graph.obj)
}
\arguments{
\item{theta}{- (sigma_e, sigma, kappa)}

\item{graph.obj}{- graph object}
}
\value{
Leave-one-out predictions
}
\description{
Prediction for the graph-Laplacian model
}
