% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expontial.covariance.R
\name{posterior.leave.stupid}
\alias{posterior.leave.stupid}
\title{prediction not using sparsity}
\usage{
posterior.leave.stupid(theta, graph.obj)
}
\arguments{
\item{theta}{- (sigma_e, sigma, kappa)}

\item{graph.obj}{- graph object}
}
\value{
Leave-one-out predictions
}
\description{
prediction not using sparsity
}
