% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expontial.covariance.R
\name{posterior.leave.v2}
\alias{posterior.leave.v2}
\title{Prediction for alpha=1 assuming observations at vertices}
\usage{
posterior.leave.v2(theta, graph.obj)
}
\arguments{
\item{theta}{- (sigma_e, sigma, kappa)}

\item{graph.obj}{- graph object}
}
\value{
Leave-one-out predictions
}
\description{
Prediction for alpha=1 assuming observations at vertices
}
