% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spde_posteriors.R
\name{posterior.mean.exp}
\alias{posterior.mean.exp}
\title{Computes the posterior expectation for each node in the graph}
\usage{
posterior.mean.exp(theta, graph, rem.edge = FALSE)
}
\arguments{
\item{theta}{- (sigma_e, sigma, kappa)}

\item{graph}{- metric_graph object}

\item{rem.edge}{- remove edge}
}
\description{
Computes the posterior expectation for each node in the graph
}
