% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spde_posteriors.R
\name{posterior.mean.matern2}
\alias{posterior.mean.matern2}
\title{Computes the posterior mean for alpha = 2}
\usage{
posterior.mean.matern2(theta, graph, rem.edge = NULL)
}
\arguments{
\item{theta}{parameters (sigma_e, sigma, kappa)}

\item{graph}{metric_graph object}
}
\description{
Computes the posterior mean for alpha = 2
}
