% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spde_posteriors.R
\name{posterior.mean.obs.matern2}
\alias{posterior.mean.obs.matern2}
\title{Computes the posterior mean for each observation in the graph based
on the alpha=2 model}
\usage{
posterior.mean.obs.matern2(theta, graph, leave.edge.out = FALSE)
}
\arguments{
\item{theta}{parameters (sigma_e, sigma, kappa)}

\item{graph}{metric_graph object}

\item{leave.edge.out}{compute the expectation of the graph if the
observatrions are not on the edge}
}
\description{
Computes the posterior mean for each observation in the graph based
on the alpha=2 model
}
