% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_inference.R
\name{posterior_crossvalidation_covariance}
\alias{posterior_crossvalidation_covariance}
\title{Leave-one-out crossvalidation for \code{graph_lme} models assuming observations at
the vertices of metric graphs.}
\usage{
posterior_crossvalidation_covariance(object)
}
\arguments{
\item{object}{A fitted model using the \code{graph_lme()} function.}
}
\value{
Vector with all predictions
}
\description{
Leave-one-out crossvalidation for \code{graph_lme} models assuming observations at
the vertices of metric graphs.
}
\details{
This function does not use sparsity for any model.
}
