% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spde_posteriors.R
\name{posterior_mean_alpha1}
\alias{posterior_mean_alpha1}
\title{Computes the posterior expectation for each node in the graph}
\usage{
posterior_mean_alpha1(theta, graph, rem.edge = FALSE)
}
\arguments{
\item{theta}{\itemize{
\item (sigma_e, sigma, kappa)
}}

\item{graph}{\itemize{
\item metric_graph object
}}

\item{rem.edge}{\itemize{
\item remove edge
}}
}
\description{
Computes the posterior expectation for each node in the graph
}
