% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_inference.R
\name{posterior_mean_covariance}
\alias{posterior_mean_covariance}
\title{Posterior mean for models assuming observations at vertices}
\usage{
posterior_mean_covariance(theta, graph, model = "alpha1")
}
\arguments{
\item{theta}{estimated model parameters (sigma_e, sigma, kappa)}

\item{graph}{metric graph  object}

\item{model}{Type of model: "alpha1" gives SPDE with alpha=1, "GL" gives
graph Laplacian, and "isoExp" gives isotropic exponential}

\item{type}{if 'obs' the posterior mean is computed at the observation
locations, if 'vertices' it is computed at all vertices.}
}
\value{
Vector with the posterior mean evaluated at the observation locations
}
\description{
Posterior mean for models assuming observations at vertices
}
\details{
This function does not use sparsity for any model.
}
