% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spde_posteriors.R
\name{posterior_mean_obs_alpha1}
\alias{posterior_mean_obs_alpha1}
\title{Computes the posterior mean for the alpha=1 model}
\usage{
posterior_mean_obs_alpha1(theta, graph, type = "obs", leave.edge.out = FALSE)
}
\arguments{
\item{theta}{parameters (sigma_e, sigma, kappa)}

\item{graph}{metric_graph object}

\item{type}{decides where to predict, 'obs' or 'mesh'.}

\item{leave.edge.out}{compute the mean of the graph if the observations
are not on the edge}
}
\description{
Computes the posterior mean for the alpha=1 model
}
