% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spde_posteriors.R
\name{posterior_mean_obs_alpha2}
\alias{posterior_mean_obs_alpha2}
\title{Computes the posterior mean for the alpha=2 model}
\usage{
posterior_mean_obs_alpha2(theta, graph, type = "obs", leave.edge.out = FALSE)
}
\arguments{
\item{theta}{parameters (sigma_e, sigma, kappa)}

\item{graph}{metric_graph object}

\item{type}{Set to 'obs' for computation at observation locations, or to
'mesh' for computation at mesh locations.}

\item{leave.edge.out}{compute the expectation of the graph if the}
}
\description{
Computes the posterior mean for the alpha=2 model
}
