% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance_alpha1.R
\name{precision_exp_line}
\alias{precision_exp_line}
\title{Precision matrix for exponential covariance}
\usage{
precision_exp_line(kappa, sigma, t, t_sorted = FALSE)
}
\arguments{
\item{kappa}{parameter kappa}

\item{sigma}{parameter sigma}

\item{t}{(n x 1) position on the line}

\item{t_sorted}{(bool)}
}
\description{
Computes the precision matrix of observations on an interval
for a Gaussian process with an exponential covariance
\deqn{r(h) = \sigma^2\exp(-\kappa h)/(2\kappa)}{r(h) = sigma^2*(exp(-kappa*h)/(2*kappa)}
}
