% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{profile.graph_lme}
\alias{profile.graph_lme}
\alias{profile}
\title{Profile method for \code{graph_lme} objects}
\usage{
\method{profile}{graph_lme}(
  fitted,
  which_par = NULL,
  alphamax = 0.01,
  maxpts = 100,
  delta = NULL,
  delta_cutoff = 1/8,
  maxmult = 10,
  minstep = 1e-06,
  verbose = FALSE,
  optim_method = "L-BFGS-B",
  parallel = FALSE,
  n_cores = parallel::detectCores() - 1,
  optim_controls = list()
)
}
\arguments{
\item{fitted}{A fitted model using \code{graph_lme()}.}

\item{alphamax}{a number in \verb{(0,1)} such that \code{1 - alphamax} is the maximum alpha value for likelihood ratio confidence regions; used to establish the range of values to be profiled.}

\item{maxpts}{maximum number of points (in each direction, for each parameter) to evaluate in attempting to construct the profile.}

\item{delta}{stepping scale for deciding on next point to profile.}

\item{maxmult}{maximum multiplier of the original step size allowed, defaults to 10.}

\item{verbose}{Should update messages be printed with updates?}

\item{optim_method}{Which optimization method to pass to \code{optim}.}

\item{parallel}{Should parallel optimization be used?}

\item{n_cores}{Number of cores to be used in case of parallel optimization.}

\item{optim_controls}{Additional controls to be passed to \code{optim}}

\item{which}{A character vector indicating which parameters to profile. \code{NULL} means all parameters. \code{.fixed} will act as a vector of all fixed effects, \code{.random} will act as a vector of all random effects.}

\item{delta.cutoff}{stepping scale (see delta) expressed as a fraction of the target maximum value of the profile on the square-root-deviance scale.}
}
\value{
A \code{data.frame} containing the profiled parameters.
}
\description{
Function providing a summary of several informations/characteristics of a metric graph object.
}
