% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psp_to_metricgraph.R
\name{psp.to.graph}
\alias{psp.to.graph}
\title{function to export a psp object, from spatstat package, to metric graph object}
\usage{
psp.to.graph(psp.object)
}
\description{
function to export a psp object, from spatstat package, to metric graph object
}
