% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.prec.R
\name{Q.alpha2.base}
\alias{Q.alpha2.base}
\title{Title precision for alpha=2 process on interval}
\usage{
Q.alpha2.base(l, kappa, tau)
}
\arguments{
\item{l}{length of interval}

\item{kappa}{range parameter}

\item{tau}{precision parameter}
}
\value{
precision matrix
}
\description{
Title precision for alpha=2 process on interval
}
