% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.prec.R
\name{Q.alpha2.line}
\alias{Q.alpha2.line}
\title{Build components for Q on line for alpha=2}
\usage{
Q.alpha2.line(loc, kappa, tau)
}
\arguments{
\item{loc}{locations}

\item{kappa}{range parameter}

\item{tau}{precision parameter}
}
\value{
List with elements for the precision
}
\description{
Build components for Q on line for alpha=2
}
