% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expontial.covariance.R
\name{Q.exp.line}
\alias{Q.exp.line}
\title{Compute the precision matrix of observations one a line}
\usage{
Q.exp.line(theta, t, t_sorted = FALSE)
}
\arguments{
\item{theta}{- kappa}

\item{t}{- (n x 1) relative position on the line start with 0 end with 1}

\item{t_sorted}{- (bool)}

\item{l_e}{- (double) length of the line}
}
\description{
Compute the precision matrix of observations one a line
}
