% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matern2.covariance.R
\name{Q.matern2}
\alias{Q.matern2}
\title{Building the precision matrix for all the edges in the matern alpha=2 case
The ordering is acording to EtV where for each edge there are four random variables
processes and derivate for lower and upper edge end point}
\usage{
Q.matern2(theta, V, EtV, El, BC = 1, build = T)
}
\arguments{
\item{theta}{- kappa, sigma}

\item{V}{vertex position}

\item{EtV}{[,2-3] index of upper and lower edge}

\item{El}{length of each vertex}

\item{BC}{boundary condition =0 neumann, 1 = correct boundary to stationary}

\item{build}{(bool) if true return Q the precision matrix otherwise return list(i,j,x, nv)}
}
\value{
Q (precision matrix)
}
\description{
Building the precision matrix for all the edges in the matern alpha=2 case
The ordering is acording to EtV where for each edge there are four random variables
processes and derivate for lower and upper edge end point
}
