% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.prec.R
\name{Q.vertex.exp}
\alias{Q.vertex.exp}
\title{computes the precision matrix for the vertices for an exponential graph model}
\usage{
Q.vertex.exp(P, E, kappa, l = NULL, adjust = FALSE)
}
\arguments{
\item{P}{(n x 2) Vertex locations}

\item{E}{(m x 2) Edges}

\item{kappa}{matern parameter}
}
\description{
computes the precision matrix for the vertices for an exponential graph model
}
