% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spde_precision.R
\name{Qalpha1_v2}
\alias{Qalpha1_v2}
\title{The precision matrix for all vertices in the alpha=1 case}
\usage{
Qalpha1_v2(theta, graph, w = 0.5, BC = 0, build = TRUE)
}
\arguments{
\item{theta}{\itemize{
\item sigma, kappa
}}

\item{graph}{metric_graph object}

\item{w}{(\link{0,1}) how two weight the top edge}

\item{BC}{boundary conditions for degree=1 vertices. BC =0 gives Neumann
boundary conditions and BC=1 gives stationary boundary conditions}

\item{build}{(bool) if TRUE return the precision matrix otherwise return
a list(i,j,x, nv)}
}
\value{
Precision matrix or list
}
\description{
The precision matrix for all vertices in the alpha=1 case
}
