% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spde_precision.R
\name{Qalpha2}
\alias{Qalpha2}
\title{The precision matrix for all vertices in the alpha=2 case}
\usage{
Qalpha2(theta, graph, BC = 1, build = TRUE)
}
\arguments{
\item{theta}{\itemize{
\item sigma, kappa
}}

\item{graph}{metric_graph object}

\item{BC}{boundary conditions for degree=1 vertices. BC =0 gives Neumann
boundary conditions and BC=1 gives stationary boundary conditions}

\item{build}{(bool) if TRUE return the precision matrix otherwise return
a list(i,j,x, nv)}
}
\value{
Precision matrix or list
}
\description{
The precision matrix for all vertices in the alpha=2 case
}
\details{
This is the unconstrained precision matrix of the process and its
derivatives. The ordering of the variables is acording to graph$E, where for
each edge there are four random variables: processes and derivate for
lower and upper edge end points
}
