% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expontial.covariance.R
\name{r_1}
\alias{r_1}
\title{The exponential covariance}
\usage{
r_1(D, theta)
}
\arguments{
\item{D}{- vector, or matrix distance}

\item{theta}{- kappa, sigma}
}
\description{
The exponential covariance
}
