% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spde_precision.R
\name{r_1_circle}
\alias{r_1_circle}
\title{the Whittle--Matern covariance with alpha=1 on a circle}
\usage{
r_1_circle(t, l_e, theta)
}
\arguments{
\item{t}{locations}

\item{l_e}{circle perimeter}

\item{theta}{parameters kappa and sigma}
}
\description{
the Whittle--Matern covariance with alpha=1 on a circle
}
