% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matern2.covariance.R
\name{r_2}
\alias{r_2}
\title{The Matern alpha=2 covariance}
\usage{
r_2(D, theta, deriv = 0)
}
\arguments{
\item{D}{- vector, or matrix distance}

\item{theta}{- kappa, sigma}

\item{deriv}{- (0,1,2) no derivative, first, or second}
}
\description{
The Matern alpha=2 covariance
}
