% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expontial.covariance.R
\name{sample.line.exponential}
\alias{sample.line.exponential}
\title{sample a line given end points}
\usage{
sample.line.exponential(
  theta,
  u_e,
  l_e,
  t = NULL,
  Line = NULL,
  nt = 100,
  py = NULL,
  y = NULL,
  sample = TRUE
)
}
\arguments{
\item{theta}{(3 x 1) sigma_e, sigma, kappa}

\item{u_e}{(2 x 1) the two end points}

\item{l_e}{(1 x 1) line length}

\item{t}{(n x 1) distance on the line to be sampled from (not end points)}

\item{Line}{Spatial.dataframe}

\item{nt}{(1 x 1) number of equidistance points to sample from if t is  null}

\item{py}{(n x 1) observation locations}

\item{y}{(n x 1) observations}

\item{sample}{(bool) if true sample else return mean}
}
\description{
sample a line given end points
}
