% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matern2.covariance.R
\name{sample.line.matern2}
\alias{sample.line.matern2}
\title{sample a line given end points}
\usage{
sample.line.matern2(
  theta,
  u_e,
  l_e,
  t = NULL,
  Line = NULL,
  nt = 100,
  py = NULL,
  y = NULL,
  sample = TRUE
)
}
\arguments{
\item{theta}{- (3 x 1) sigma_e, sigma, kappa}

\item{u_e}{- (4 x 1) X(0),X'(0), X(l_e),X'(l_e) the two end points and there deriv}

\item{l_e}{- (1 x 1) line length}

\item{t}{- (n x 1) distance on the line to be sampled from (not end points)}

\item{Line}{- Spatial.datafram}

\item{nt}{- (1 x 1) number of equidistance points to sample from if t is  null}

\item{py}{- (n x 1) observation location}

\item{y}{- (n x 1) observations}

\item{sample}{- (bool) if true sample else return mean}
}
\description{
sample a line given end points
}
