% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spde_sampling.R
\name{sample_spde}
\alias{sample_spde}
\title{Sample Whittle-Matérn field on graph}
\usage{
sample_spde(
  kappa,
  sigma,
  sigma_e = 0,
  alpha = 1,
  graph,
  PtE = NULL,
  type = "manual",
  posterior = FALSE,
  nsim = 1,
  method = c("conditional", "Q"),
  BC = 1
)
}
\arguments{
\item{kappa}{parameter kappa}

\item{sigma}{parameter sigma}

\item{sigma_e}{parameter sigma_e}

\item{graph}{metric_graph object}

\item{PtE}{matrix with locations (edge, normalized distance on edge) to
sample from}

\item{type}{If manual is set, then sampling is done at the locations
specified in PtE. Set to "mesh" for simulation at mesh nodes, and to "obs"
for simulation at observation locations.}

\item{posterior}{sample conditionally on the observations?}

\item{nsim}{number of samples to be generated.}

\item{method}{Which method to sample? The options are "conditional" and "Q". "Q" is more stable but takes longer.}

\item{BC}{boundary conditions for degree=1 vertices. BC =0 gives Neumann
boundary conditions and BC=1 gives stationary boundary conditions}
}
\value{
sample evaluated at the mesh nodes in the graph
}
\description{
Sample Whittle-Matérn field on graph
}
\details{
Samples a Gaussian Whittle-Matérn field on a graph, either from the prior
or conditionally on observations
\deqn{y_i = u(t_i) + sigma_e e_i}{y_i = u(t_i) + sigma_e e_i}
in the graph,  where \eqn{e_i} are independent standard Gaussian variables.
}
