% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spde_sampling.R
\name{sample_spde_mesh}
\alias{sample_spde_mesh}
\title{Sample Whittle-Matérn field on graph}
\usage{
sample_spde_mesh(kappa, sigma, sigma_e, alpha = 1, graph, posterior = FALSE)
}
\arguments{
\item{kappa}{parameter kappa}

\item{sigma_e}{parameter sigma_e}

\item{posterior}{sample conditionally on the observations?}

\item{theta}{parameter theta}

\item{u}{sample evaluated at the mesh nodes in the graph}
}
\description{
Sample Whittle-Matérn field on graph
}
\details{
Samples a Gaussian Whittle-Matérn field on a graph, either from the prior
or conditionally on observations
\deqn{y_i = u(t_i) + sigma_e e_i}{y_i = u(t_i) + sigma_e e_i}
in the graph,  where \eqn{e_i} are independent standard Gaussian variables.
}
