% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selected_inv.R
\name{selected_inv}
\alias{selected_inv}
\title{Selected Inverse Calculation}
\usage{
selected_inv(Q)
}
\arguments{
\item{Q}{A sparse matrix in dgCMatrix format}
}
\value{
A numeric vector containing the selected inverse
}
\description{
Selected Inverse Calculation
}
