% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spCircle.R
\name{spCircle}
\alias{spCircle}
\title{Circular \code{sp} SpatialPolygons}
\usage{
spCircle(
  radius,
  spUnits = CRS(projargs = as.character(NA)),
  centerPoint = c(x = 0, y = 0),
  nptsPerimeter = 100,
  spID = paste("circle", round(1000 * stats::runif(1)), sep = ":"),
  ...
)
}
\arguments{
\item{radius}{the radius of the circle}

\item{spUnits}{the CRS units}

\item{centerPoint}{the circle's cente}

\item{nptsPerimeter}{the number of points forming the perimeter of the polygon}

\item{spID}{the spatial ID for the object}

\item{...}{additional arguments to be passed.}
}
\value{
a list with...
spCircle = the SpatialPolygons circular object
location =  the SpatialPoints center point
}
\description{
This routine will create an "sp" circular SpatialPolygons object based on
}
