% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spde_posteriors.R
\name{spde_posterior_mean}
\alias{spde_posterior_mean}
\title{Computes the posterior expectation for SPDE models}
\usage{
spde_posterior_mean(
  theta,
  graph,
  alpha = 1,
  type = "mesh",
  leave_edge_out = FALSE
)
}
\arguments{
\item{theta}{parameters (sigma_e, sigma, kappa)}

\item{graph}{metric_graph object}

\item{alpha}{alpha parameter (1 or 2)}

\item{leave_edge_out}{if obs=TRUE, then leave_edge_out = TRUE means that the
posterior mean is computed for each observation based on all observations
which are not on that edge. If obs=FALSE, then leave_edge_out = e means that
the posterior mean is computed based on all observations except those on that
particular edge.}

\item{obs}{if TRUE, then the posterior mean is calculated at the observation
locations, otherwise it is computed at the vertices in the graph}
}
\description{
Computes the posterior expectation for SPDE models
}
