% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_lme.R
\name{summary.graph_lme}
\alias{summary.graph_lme}
\title{Summary Method for \code{graph_lme} Objects}
\usage{
\method{summary}{graph_lme}(object, all_times = FALSE, ...)
}
\arguments{
\item{object}{an object of class \code{graph_lme} containing results from the
fitted model.}

\item{all_times}{Show all computed times.}

\item{...}{not used.}
}
\value{
An object of class \code{summary_graph_lme} containing information
about a \emph{graph_lme} object.
}
\description{
Function providing a summary of results related to metric graph
mixed effects regression models.
}
