% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{summary.metric_graph}
\alias{summary.metric_graph}
\title{Summary Method for \code{metric_graph} Objects}
\usage{
\method{summary}{metric_graph}(
  object,
  messages = FALSE,
  compute_characteristics = TRUE,
  check_euclidean = TRUE,
  check_distance_consistency = TRUE,
  ...
)
}
\arguments{
\item{object}{an object of class \code{metric_graph}.}

\item{messages}{Should message explaining how to build the results be given for missing quantities?}

\item{compute_characteristics}{Should the characteristics of the graph be computed?}

\item{check_euclidean}{Check if the graph has Euclidean edges?}

\item{check_distance_consistency}{Check the distance consistency assumption?#'}

\item{...}{not used.}
}
\value{
An object of class \code{summary_graph_lme} containing information
about a \emph{metric_graph} object.
}
\description{
Function providing a summary of several informations/characteristics of a metric graph object.
}
