% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_spde.R
\name{summary.metric_graph_spde_result}
\alias{summary.metric_graph_spde_result}
\title{Summary for posteriors of field parameters for an \code{inla_rspde}
model from a \code{rspde.result} object}
\usage{
\method{summary}{metric_graph_spde_result}(object, digits = 6, ...)
}
\arguments{
\item{object}{A \code{rspde.result} object.}

\item{digits}{Integer, used for number formatting with signif()}

\item{...}{Currently not used.}
}
\value{
A \code{data.frame} containing the summary.
}
\description{
Summary for posteriors of 'rSPDE' field parameters in
their original scales.
}
