% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.object.R
\name{vertex.to.line}
\alias{vertex.to.line}
\title{Convert line object to graph information}
\usage{
vertex.to.line(Lines)
}
\arguments{
\item{Lines}{object}
}
\value{
A list with elements V, EtV, edge_lengths
}
\description{
Convert line object to graph information
}
