% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_spde.R
\name{bru_mapper.inla_metric_graph_spde}
\alias{bru_mapper.inla_metric_graph_spde}
\alias{bru_get_mapper.inla_metric_graph_spde}
\alias{ibm_n.bru_mapper_inla_metric_graph_spde}
\alias{ibm_values.bru_mapper_inla_metric_graph_spde}
\alias{ibm_jacobian.bru_mapper_inla_metric_graph_spde}
\title{Metric graph 'inlabru' mapper}
\usage{
bru_get_mapper.inla_metric_graph_spde(model, ...)

ibm_n.bru_mapper_inla_metric_graph_spde(mapper, ...)

ibm_values.bru_mapper_inla_metric_graph_spde(mapper, ...)

ibm_jacobian.bru_mapper_inla_metric_graph_spde(mapper, input, ...)
}
\arguments{
\item{model}{An \code{inla_metric_graph_spde} for which to construct or extract a mapper}

\item{\dots}{Arguments passed on to other methods}

\item{mapper}{A \code{bru_mapper.inla_metric_graph_spde} object}

\item{input}{The values for which to produce a mapping matrix}
}
\description{
Metric graph 'inlabru' mapper
}
