% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance_alpha1.R
\name{covariance_alpha1_mesh}
\alias{covariance_alpha1_mesh}
\title{Compute covariance of a point to the mesh points of the graph for
alpha=1 model}
\usage{
covariance_alpha1_mesh(P, kappa, tau, range, sigma, graph, scale = FALSE)
}
\arguments{
\item{P}{(2 x 1) point with edge number and normalized location on the edge}

\item{kappa}{parameter kappa from the SPDE}

\item{tau}{parameter tau from the SPDE}

\item{range}{range parameter}

\item{sigma}{std. deviation parameter}

\item{graph}{metric_graph object}

\item{scale}{scale the covariance by \code{2*kappa} so that sigma corresponds to
the marginal standard deviation (default FALSE)}
}
\value{
vector with covariance values (order Vertex of Graph then mesh$PtE)
}
\description{
Compute covariance of a point to the mesh points of the graph for
alpha=1 model
}
