% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.object.R
\name{graph.obj}
\alias{graph.obj}
\title{A general graph objects,
builds an object of sp::Lines object each Lines is assumes to be an edge.
Thus graphs can only be connected by end points}
\description{
A general graph objects,
builds an object of sp::Lines object each Lines is assumes to be an edge.
Thus graphs can only be connected by end points

A general graph objects,
builds an object of sp::Lines object each Lines is assumes to be an edge.
Thus graphs can only be connected by end points
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{El}}{length of edges}

\item{\code{EID}}{ID of edges}

\item{\code{A}}{constraint matrix, just to setup Kirchoff constraint}

\item{\code{CBobj}}{svd stuct obj}

\item{\code{Points}}{Observations in SpatialPointsDataFrame}

\item{\code{y}}{the data connected to P}

\item{\code{PtE}}{Points to Line (connected to Points),
[,1] - edge index,
[,2] - distance along the line (i.e. distance to initial point)}

\item{\code{V}}{poisition in the space [,1] - id [,-1] - point}

\item{\code{EtV}}{[,1]- index of Lines, [,2] - vertex lower edge, [,3] - vertex upper edge}

\item{\code{nE}}{number of edges}

\item{\code{Lines}}{List of Lines object for building the graph}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GPGraph::graph-new}{\code{graph.obj$new()}}
\item \href{#method-GPGraph::graph-add_observations}{\code{graph.obj$add_observations()}}
\item \href{#method-GPGraph::graph-add_observations2}{\code{graph.obj$add_observations2()}}
\item \href{#method-GPGraph::graph-buildA}{\code{graph.obj$buildA()}}
\item \href{#method-GPGraph::graph-clone}{\code{graph.obj$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GPGraph::graph-new"></a>}}
\if{latex}{\out{\hypertarget{method-GPGraph::graph-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{graph.obj$new(Lines.in = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Lines.in}}{sp object SpatialLinesDataFrame or SpatialLines

add observations to the object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GPGraph::graph-add_observations"></a>}}
\if{latex}{\out{\hypertarget{method-GPGraph::graph-add_observations}{}}}
\subsection{Method \code{add_observations()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{graph.obj$add_observations(Spoints, y = NULL, y.index = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Spoints}}{SpatialPoints or SpatialPointsDataFrame of the observations}

\item{\code{y}}{(n x 1) the value of the observations}

\item{\code{y.index}}{(string, int) position in Spoints where y is located
add observations to the object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GPGraph::graph-add_observations2"></a>}}
\if{latex}{\out{\hypertarget{method-GPGraph::graph-add_observations2}{}}}
\subsection{Method \code{add_observations2()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{graph.obj$add_observations2(y, PtE, Spoints = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{y}}{(n x 1) the value of the observations}

\item{\code{PtE}}{(n x 2) edge index, distance on index
build Kirchoff constraint matrix from edges, NOT implemented for circles (i.e. self closed edges)}

\item{\code{Spoints}}{SpatialPoints or SpatialPointsDataFrame of the observations}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GPGraph::graph-buildA"></a>}}
\if{latex}{\out{\hypertarget{method-GPGraph::graph-buildA}{}}}
\subsection{Method \code{buildA()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{graph.obj$buildA(alpha, edge_constraint = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{alpha}}{(int) which type of constraint (currently only 2 implemented)}

\item{\code{edge_constraint}}{(bool) if true add constraint on vertices of degree 1.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GPGraph::graph-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GPGraph::graph-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{graph.obj$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
