% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.object.R
\name{graph.obj}
\alias{graph.obj}
\title{A general graph objects,
builds an object of sp::Lines object each Lines is assumes to be an edge.
Thus graphs can only be connected by end points}
\description{
A general graph objects,
builds an object of sp::Lines object each Lines is assumes to be an edge.
Thus graphs can only be connected by end points

A general graph objects,
builds an object of sp::Lines object each Lines is assumes to be an edge.
Thus graphs can only be connected by end points
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{El}}{length of edges}

\item{\code{V}}{poisition in the space}

\item{\code{EtV}}{[,1]- index of Lines, [,2] - vertex lower edge, [,3] - vertex upper edge}

\item{\code{nE}}{number of edges}

\item{\code{Lines}}{List of Lines object for building the graph}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GPGraph::graph-new}{\code{graph.obj$new()}}
\item \href{#method-GPGraph::graph-get_name}{\code{graph.obj$get_name()}}
\item \href{#method-GPGraph::graph-clone}{\code{graph.obj$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GPGraph::graph-new"></a>}}
\if{latex}{\out{\hypertarget{method-GPGraph::graph-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{graph.obj$new(Lines.in = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Lines}}{sp object SpatialLinesDataFrame or list(Lines)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GPGraph::graph-get_name"></a>}}
\if{latex}{\out{\hypertarget{method-GPGraph::graph-get_name}{}}}
\subsection{Method \code{get_name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{graph.obj$get_name()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GPGraph::graph-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GPGraph::graph-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{graph.obj$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
