% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{graph_components}
\alias{graph_components}
\title{Create metric graphs for connected components of a SpatialLines object}
\usage{
graph_components(lines, by_length = TRUE, only_largest = FALSE, ...)
}
\arguments{
\item{lines}{Object of class \code{SpatialLines}}

\item{by_length}{Sort the components by total edge length? If FALSE,
the components are sorted by the number of vertices.}

\item{only_largest}{if TRUE, only return the largest component.
Otherwise return an ordered list with the components (largest first)}

\item{...}{additional arguments used when specifying the graph}
}
\value{
A \code{metric_graph} object created from the largest component, or a
list of \code{metric_graph} objects for all connected components
}
\description{
Create metric graphs for connected components of a SpatialLines object
}
\examples{
library(sp)
line1 <- Line(rbind(c(0, 0), c(1, 0)))
line2 <- Line(rbind(c(1, 0), c(2, 0)))
line3 <- Line(rbind(c(1, 1), c(2, 1)))
Lines <-  SpatialLines(list(Lines(list(line1), ID = "1"),
                           Lines(list(line2), ID = "2"),
                           Lines(list(line3), ID = "3")))
graphs <- graph_components(Lines, only_largest = FALSE)
p <- graphs[[1]]$plot(edge_color = "red")
graphs[[2]]$plot(p = p, edge_color = "blue")
}
