% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_lgcp.R
\name{graph_lgcp}
\alias{graph_lgcp}
\title{Simulation of log-Gaussian Cox process on metric graph}
\usage{
graph_lgcp(kappa, tau, alpha, graph)
}
\arguments{
\item{kappa}{range parameter kappa}

\item{tau}{variance parameter}

\item{alpha}{smoothness parameter (1 or 2)}

\item{graph}{metric_graph object}
}
\value{
List with Gaussian process sample and simulated points
}
\description{
Simulation of log-Gaussian Cox process on metric graph
}
