% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_lgcp.R
\name{graph_lgcp}
\alias{graph_lgcp}
\title{Simulation of log-Gaussian Cox processes driven by Whittle-Matérn
fields on metric graphs}
\usage{
graph_lgcp(intercept = 0, sigma, range, alpha, graph)
}
\arguments{
\item{intercept}{Mean value of the Gaussian process.}

\item{sigma}{Parameter for marginal standard deviations.}

\item{range}{Parameter for practical correlation range.}

\item{alpha}{Smoothness parameter (1 or 2).}

\item{graph}{A \code{metric_graph} object.}
}
\value{
List with Gaussian process sample and simulated points.
}
\description{
Simulation of log-Gaussian Cox processes driven by Whittle-Matérn
fields on metric graphs
}
