% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_lme.R
\name{graph_lme}
\alias{graph_lme}
\title{Metric graph linear mixed effects models}
\usage{
graph_lme(
  formula,
  graph,
  model = list(type = "linearModel"),
  which_repl = NULL,
  optim_method = "L-BFGS-B",
  starting_values_latent = NULL,
  start_sigma_e = NULL,
  parameterization_latent = c("matern", "spde"),
  BC = 1,
  model_matrix = TRUE,
  parallel = FALSE,
  n_cores = parallel::detectCores() - 1,
  optim_controls = list(),
  improve_hessian = FALSE,
  hessian_args = list()
)
}
\arguments{
\item{formula}{Formula object describing the relation between the response variables and the fixed effects.}

\item{graph}{A \code{metric_graph} object.}

\item{model}{The random effects model that will be used (it also includes the option of not having any random effects). It can be either a character, whose options are 'lm', for linear models without random effects; 'alpha1' and 'alpha2' for Whittle-Matern models with \eqn{\alpha}=1 and 2, respectively; 'isoExp' for a model with isotropic exponential covariance; 'GL1' and 'GL2' for a SPDE model based on graph Laplacian with \eqn{\alpha} = 1 and 2, respectively. There is also the option to provide it as a list containing the elements \code{type}, which can be \code{linearModel}, \code{WhittleMatern}, \code{graphLaplacian} or \code{isoCov}.
\code{linearModel} corresponds to a linear model without random effects.
For \code{Whittle-Matern} and \code{graph-Laplacian} models, the list must also contain a parameter \code{alpha} (which is 1 by default). For \code{isoCov} models, the list must
contain a parameter \code{cov_function}, containing the covariance function. The function accepts a string input for the following covariance functions: 'exp_covariance', 'alpha1', 'alpha2', 'GL1', 'GL2'. For another covariance function, the function itself must be provided as the \code{cov_function} argument. The default is 'exp_covariance', the
exponential covariance. We also have covariance-based versions of the Whittle-Matern and graph Laplacian models, however they are much slower, they are the following (string) values for 'cov_function': 'alpha1' and 'alpha2' for Whittle-Matern fields, and 'GL1' and 'GL2' for graph Laplacian models. Finally, for \code{Whittle-Matern} models, there is an additional parameter
\code{version}, which can be either 1 or 2, to tell which version of the likelihood should be used. Version is 1 by default.}

\item{which_repl}{vector or list containing which replicates to consider in the model.
If \code{NULL} all replicates will be considered.}

\item{optim_method}{The method to be used with \code{optim} function.}

\item{starting_values_latent}{A vector containing the starting values for the latent model. If the latent model is \code{WhittleMatern} or \code{graphLaplacian}, then the starting values should be provided as a vector of the form c(sigma,kappa) or c(sigma,range) depending on the parameterization. If the model is \code{isoCov}, then the starting values should be provided as a vector containing the parameters of the covariance function.}

\item{start_sigma_e}{Starting value for the standard deviation of the measurament error.}

\item{parameterization_latent}{The parameterization for \code{WhittleMatern} and \code{graphLaplacian} models. The options are 'matern' and 'spde'. The 'matern' parameterizes as 'sigma' and 'range', whereas the 'spde' parameterization is given in terms of 'sigma' and 'kappa'.}

\item{BC}{For \code{WhittleMatern} models. Which boundary condition to use (0,1) 0 is no adjustment on boundary point
1 is making the boundary condition stationary.}

\item{model_matrix}{logical indicating whether the model matrix should be returned as component of the returned value.}

\item{parallel}{logical. Indicating whether to use optimParallel or not.}

\item{n_cores}{Number of cores to be used if parallel is true.}

\item{optim_controls}{Additional controls to be passed to \code{optim} or \code{optimParallel}.}

\item{improve_hessian}{Should a more precise estimate of the hessian be obtained? Turning on might increase the overall time.}

\item{hessian_args}{List of controls to be used if \code{improve_hessian} is \code{TRUE}. The list can contain the arguments to be passed to the \code{method.args} argument in the \code{hessian} function. See the help of the \code{hessian} function in \code{numDeriv} package for details. Observet that it only accepts the "Richardson" method for now, the method "complex" is not supported.}
}
\value{
A list containing the fitted model.
}
\description{
Fitting linear mixed effects model in metric graphs. The random effects can be
Whittle-Matern fields on metric graphs, Whittle-Matern fields based on the
graph Laplacian as well as fields with isotropic covariance structure.
}
