% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_spde.R
\name{graph_spde}
\alias{graph_spde}
\title{INLA implementation of Whittle-Matérn fields for metric graphs.}
\usage{
graph_spde(
  graph_object,
  alpha = 1,
  stationary_endpoints = "all",
  parameterization = c("matern", "spde"),
  start_range = NULL,
  prior_range = NULL,
  start_kappa = NULL,
  start_sigma = NULL,
  prior_kappa = NULL,
  prior_sigma = NULL,
  debug = FALSE
)
}
\arguments{
\item{graph_object}{A \code{metric_graph} object.}

\item{alpha}{The order of the SPDE.}

\item{stationary_endpoints}{Which vertices of degree 1 should contain stationary boundary conditions?}

\item{parameterization}{Which parameterization to be used? The options are 'matern' (sigma and range) and 'spde' (sigma and kappa).}

\item{start_range}{Starting value for range parameter.}

\item{prior_range}{a \code{list} containing the elements \code{meanlog} and
\code{sdlog}, that is, the mean and standard deviation of the range parameter on the log scale. Will not be used if prior.kappa is non-null.}

\item{start_kappa}{Starting value for kappa.}

\item{start_sigma}{Starting value for sigma.}

\item{prior_kappa}{a \code{list} containing the elements \code{meanlog} and
\code{sdlog}, that is, the mean and standard deviation of kappa on the log scale.}

\item{prior_sigma}{a \code{list} containing the elements \code{meanlog} and
\code{sdlog}, that is, the mean and standard deviation of sigma on the log scale.}

\item{debug}{Should debug be displayed?}

\item{start_tau}{Starting value for tau.}

\item{prior_tau}{a \code{list} containing the elements \code{meanlog} and
\code{sdlog}, that is, the mean and standard deviation of the tau parameter on the log scale.}
}
\value{
An inla object.
}
\description{
This function creates an inla object that can be used
in \code{INLA} or \code{inlabru} to fit Whittle-Matérn fields
on metric graphs.
}
