% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_spde.R
\name{graph_spde_make_A}
\alias{graph_spde_make_A}
\title{Observation/prediction matrices for rSPDE models.}
\usage{
graph_spde_make_A(graph, n.repl = 1, obs_to_vert = FALSE)
}
\arguments{
\item{graph}{An object of class \code{metric_graph}}

\item{n.repl}{Number of replicates}

\item{obs_to_vert}{Should the observations be turned into vertices?}
}
\value{
The observation matrix
}
\description{
Constructs observation/prediction weight matrices
for metric graph models.
}
