% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_spde.R
\name{graph_spde_make_A}
\alias{graph_spde_make_A}
\title{Observation/prediction matrices for 'rSPDE' models}
\usage{
graph_spde_make_A(graph_spde, repl = NULL)
}
\arguments{
\item{graph_spde}{An \code{inla_metric_graph_spde} object built with the
\code{graph_spde()} function.}

\item{repl}{Which replicates? If there is no replicates, or to
use all replicates, one can set to \code{NULL}.}
}
\value{
The observation matrix.
}
\description{
Constructs observation/prediction weight matrices
for metric graph models.
}
