% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_spde.R
\name{graph_spde_make_index}
\alias{graph_spde_make_index}
\title{Model index vector generation for metric graph models}
\usage{
graph_spde_make_index(name, graph_spde, n.group = 1, n.repl = 1, ...)
}
\arguments{
\item{name}{A character string with the base name of the effect.}

\item{graph_spde}{An \code{inla_metric_graph_spde} object built with the
\code{graph_spde()} function.}

\item{n.group}{Number of groups.}

\item{n.repl}{Number of replicates.}

\item{...}{Currently not being used.}
}
\value{
A list of indexes.
}
\description{
Generates a list of named index vectors for
'INLA'-based metric graph models.
}
