% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{graph_starting_values}
\alias{graph_starting_values}
\title{Starting values for random field models on metric graphs}
\usage{
graph_starting_values(
  graph,
  model = NULL,
  data = TRUE,
  column_data = NULL,
  manual_data = NULL
)
}
\arguments{
\item{graph}{a \code{metric_graph} object.}

\item{model}{type of model, "alpha1", "alpha2", "isoExp", "GL1", and "GL2"
are supported}

\item{data}{Should the data be used to obtain improved starting values?}

\item{column_data}{The column of \code{graph$data} that contains the response variable.}

\item{manual_data}{A vector (or matrix) of response variables.}
}
\value{
A vector, \code{c(start_sigma_e, start_sigma, start_kappa)}
}
\description{
The results are given as \code{c(start_sigma_e, start_sigma, start_kappa)}
}
