% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_inference.R
\name{likelihood.exp.graph}
\alias{likelihood.exp.graph}
\title{Computes the log likelihood function fo theta for the graph object}
\usage{
likelihood.exp.graph(theta, graph.obj)
}
\arguments{
\item{theta}{- (sigma_e, sigma, kappa)}

\item{graph.obj}{- graph object}
}
\description{
Computes the log likelihood function fo theta for the graph object
}
