% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_inference.R
\name{likelihood.graph_laplacian}
\alias{likelihood.graph_laplacian}
\title{Log-likelihood calculation for graph Laplacian model}
\usage{
likelihood.graph_laplacian(theta, graph.obj)
}
\arguments{
\item{theta}{- (sigma_e, sigma, kappa)}

\item{graph.obj}{- graph object}
}
\value{
The log-likelihood
}
\description{
Log-likelihood calculation for graph Laplacian model
}
