% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_likelihoods.R
\name{likelihood_graph_covariance}
\alias{likelihood_graph_covariance}
\title{Function factory for likelihood evaluation not using sparsity}
\usage{
likelihood_graph_covariance(
  graph,
  model = "alpha1",
  y_graph,
  cov_function = NULL,
  X_cov = NULL,
  repl,
  log_scale = TRUE,
  maximize = FALSE
)
}
\arguments{
\item{graph}{metric_graph object}

\item{model}{Type of model: "alpha1" gives SPDE with alpha=1, "GL1" gives
the model based on the graph Laplacian with smoothness 1, "GL2" gives the
model based on the graph Laplacian with smoothness 2, and "isoCov" gives a
model with isotropic covariance.}

\item{y_graph}{Response vector given in the same order as the internal
locations from the graph.}

\item{cov_function}{The covariance function to be used in case 'model' is
chosen as 'isoCov'. \code{cov_function} must be a function of \verb{(h, theta_cov)},
where \code{h} is a vector, or matrix, containing the distances to evaluate the
covariance function at, and \code{theta_cov} is the vector of parameters of the
covariance function \code{cov_function}.}

\item{X_cov}{Matrix with covariates. The order must be the same as the
internal order from the graph.}

\item{repl}{Vector with the replicates to be considered. If \code{NULL} all
replicates will be considered.}

\item{log_scale}{Should the parameters \code{theta} of the returning function be
given in log-scale?}

\item{maximize}{If \code{FALSE} the function will return minus the likelihood, so
one can directly apply it to the \code{optim} function.}
}
\value{
The log-likelihood function, which is returned as a function with
parameter 'theta'. For models 'alpha1', 'alpha2', 'GL1' and 'GL2', the
parameter \code{theta} must be supplied as the vector \code{c(sigma_e, sigma, kappa)}.

For 'isoCov' model, theta must be a vector such that \code{theta[1]} is \code{sigma.e}
and the vector \code{theta[2:(q+1)]} is the input of \code{cov_function}, where \code{q} is
the number of parameters of the covariance function.

If \code{covariates} is \code{TRUE}, then the parameter \code{theta} must be supplied as the
vector \code{c(sigma_e, theta[2], ..., theta[q+1], beta[1], ..., beta[p])},
where \verb{beta[1],...,beta[p]} are the coefficients and \code{p} is the number of
covariates.

For the remaining models, if \code{covariates} is \code{TRUE}, then \code{theta} must be
supplied as the vector \code{c(sigma_e, sigma, kappa, beta[1], ..., beta[p])},
where \verb{beta[1],...,beta[p]} are the coefficients and \code{p} is the number of
covariates.
}
\description{
Function factory for likelihood evaluation not using sparsity
}
