% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_likelihoods.R
\name{likelihood_graph_covariance}
\alias{likelihood_graph_covariance}
\title{Likelihood evaluation not using sparsity}
\usage{
likelihood_graph_covariance(theta, graph, model = "alpha1")
}
\arguments{
\item{theta}{parameters (sigma_e, sigma, kappa)}

\item{graph}{metric_graph object}

\item{model}{Model for Gaussian process, supported options are alpha1 (SPDE with alpha=1),
alpha2 (SPDE with alpha=2), GL (graph Laplacian model with alpha=1), GL2 (graph Laplacian
model with alpha=2) and isoExp (model with isotropic exponential covariance)}
}
\value{
The log-likelihood
}
\description{
Likelihood evaluation not using sparsity
}
