% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_likelihoods.R
\name{likelihood_graph_laplacian}
\alias{likelihood_graph_laplacian}
\title{Function factory for likelihood evaluation for the graph Laplacian model}
\usage{
likelihood_graph_laplacian(
  graph,
  alpha,
  covariates = FALSE,
  log_scale = TRUE,
  maximize = FALSE
)
}
\arguments{
\item{graph}{metric_graph object}

\item{alpha}{integer greater or equal to 1. Order of the equation.}

\item{covariates}{Logical. If \code{TRUE}, the model will be considered with covariates. It requires \code{graph} to have
covariates included by the method \code{add_covariates()}.}

\item{log_scale}{Should the parameters \code{theta} of the returning function be given in log scale?}

\item{maximize}{If \code{FALSE} the function will return minus the likelihood, so one can directly apply it to the \code{optim} function.}
}
\value{
The log-likelihood function, which is returned as a function with parameter 'theta'.
The parameter \code{theta} must be supplied as
the vector \code{c(sigma_e, sigma, kappa)}.

If \code{covariates} is \code{TRUE}, then the parameter \code{theta} must be supplied as the vector \code{c(sigma_e, sigma, kappa, beta[1], ..., beta[p])},
where \verb{beta[1],...,beta[p]} are the coefficients and \code{p} is the number of covariates.
}
\description{
Function factory for likelihood evaluation for the graph Laplacian model
}
