% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_likelihoods.R
\name{likelihood_graph_spde}
\alias{likelihood_graph_spde}
\title{Log-likelihood calculation for alpha=1 model}
\usage{
likelihood_graph_spde(theta, graph, alpha = 1, version = 1)
}
\arguments{
\item{theta}{(sigma_e, sigma, kappa)}

\item{graph}{metric_graph object}

\item{alpha}{the order of the SPDE}

\item{version}{if 1, the likelihood is computed by integrating out
the vertex locations, if 2, no integration is done}
}
\value{
The log-likelihood
}
\description{
Log-likelihood calculation for alpha=1 model
}
