% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psp_to_metricgraph.R
\name{linnet.to.graph}
\alias{linnet.to.graph}
\title{Convert a \code{linnet} object to a metric graph object}
\usage{
linnet.to.graph(linnet.object, crs)
}
\arguments{
\item{linnet.object}{A \code{linnet} object to be converted.}
}
\value{
A metric graph object with edges defined by the network.
}
\description{
This function converts a \code{linnet} object (from the \code{spatstat} package) into a metric graph object.
}
