% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pems}
\alias{pems}
\title{Traffic speed data}
\format{
\subsection{\code{pems}}{

A list with three elements:
\describe{
\item{lines}{A \code{SpatialLines} object containing the road segments}
\item{PtE}{Locations of the observations on the road segments}
\item{Y}{Obsercations of traffic speed}
...
}
}
}
\source{
add
}
\usage{
pems
}
\description{
A subset of data from ...
}
\keyword{datasets}
