% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_spde.R
\name{plot.graph_bru_pred}
\alias{plot.graph_bru_pred}
\title{Plot of predicted values with 'inlabru'}
\usage{
\method{plot}{graph_bru_pred}(x, y = NULL, vertex_size = 0, ...)
}
\arguments{
\item{x}{A predicted object obtained with the \code{predict} method.}

\item{y}{Not used.}

\item{vertex_size}{Size of the vertices.}

\item{...}{Additional parameters to be passed to plot_function.}
}
\value{
A 'ggplot2' object.
}
\description{
Auxiliary function to obtain plots of the predictions of the field
using 'inlabru'.
}
