% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_spde.R
\name{plot.graph_bru_proc_pred}
\alias{plot.graph_bru_proc_pred}
\title{Plot of processed predicted values with 'inlabru'}
\usage{
\method{plot}{graph_bru_proc_pred}(x, y = NULL, vertex_size = 0, ...)
}
\arguments{
\item{x}{A processed predicted object obtained with the \code{process_rspde_predictions} function.}

\item{y}{Not used.}

\item{vertex_size}{Size of the vertices.}

\item{...}{Additional parameters to be passed to plot_function.}
}
\value{
A 'ggplot2' object.
}
\description{
Auxiliary function to obtain plots of the processed predictions of the field
using 'inlabru'.
}
