% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.plot.R
\name{plot_obs}
\alias{plot_obs}
\title{plot the observation}
\usage{
plot_obs(graph, yhat = NULL)
}
\arguments{
\item{graph}{(graph.obj)}

\item{yhat}{- (n x 1) prediction of the obsveration}
}
\description{
plot the observation
}
