% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance_alpha1.R
\name{plot_r_1}
\alias{plot_r_1}
\title{plot the exponential covariance for parameter set}
\usage{
plot_r_1(kappa, sigma, sigma_e, t = NULL)
}
\arguments{
\item{kappa}{parameter kappa}

\item{sigma}{parameter sigma}

\item{sigma_e}{parameter sigma_e (nugget)}
}
\description{
plot the exponential covariance for parameter set
}
