% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spde_precision.R
\name{plot_r_1}
\alias{plot_r_1}
\title{plot the exponential covariance for parameter set}
\usage{
plot_r_1(theta, t = NULL)
}
\arguments{
\item{theta}{parameters sigma_e (nugget), sigma and kappa}
}
\description{
plot the exponential covariance for parameter set
}
