% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance_alpha2.R
\name{plot_r_2}
\alias{plot_r_2}
\title{plot the Matern alpha = 2}
\usage{
plot_r_2(kappa, sigma, sigma_e, t = NULL)
}
\arguments{
\item{kappa}{parameter kappa}

\item{sigma}{parameter sigma}

\item{sigma_e}{parameter sigma_e}
}
\description{
plot the Matern alpha = 2
}
