% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expontial.covariance.R
\name{posterior.crossvalidation.covariance}
\alias{posterior.crossvalidation.covariance}
\title{Prediction for models assuming observations at vertices}
\usage{
posterior.crossvalidation.covariance(
  theta,
  graph.obj,
  model = "alpha1",
  ind = NULL
)
}
\arguments{
\item{theta}{Estimated model parameters (sigma_e, sigma, kappa)}

\item{graph.obj}{Graph object}

\item{model}{Type of model: "alpha1" gives SPDE with alpha=1, "GL" gives
graph Laplacian, and "isoExp" gives isotropic exponential}

\item{ind}{Indices for cross validation. It should be a vector of the same length as the
data, with integer values representing each group in the cross-validation.
If NULL, leave-one-out cross validation is performed.}
}
\value{
Vector with all predictions
}
\description{
Prediction for models assuming observations at vertices
}
\details{
This function does not use sparsity for any model.
}
