% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_inference.R
\name{posterior.leave.stupid}
\alias{posterior.leave.stupid}
\title{prediction not using sparsity}
\usage{
posterior.leave.stupid(theta, graph)
}
\arguments{
\item{theta}{- (sigma_e, sigma, kappa)}

\item{graph}{- metric_graph object}
}
\value{
Leave-one-out predictions
}
\description{
prediction not using sparsity
}
