% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matern2.covariance.R
\name{posterior.mean.matern2}
\alias{posterior.mean.matern2}
\title{Computes the posterior mean}
\usage{
posterior.mean.matern2(theta, graph.obj, rem.edge = NULL)
}
\arguments{
\item{theta}{- (sigma_e, sigma, kappa)}

\item{graph.obj}{- graphical object}
}
\description{
Computes the posterior mean
}
