% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expontial.covariance.R
\name{posterior.mean.obs.exp}
\alias{posterior.mean.obs.exp}
\title{Computes the posterior expectation for each observation in the graph}
\usage{
posterior.mean.obs.exp(theta, graph.obj, leave.edge.out = F)
}
\arguments{
\item{theta}{- (sigma_e, sigma, kappa)}

\item{graph.obj}{- graph object}

\item{leave.edge.out}{- compute the expectation of the graph if the observatrions are not on the edge}
}
\description{
Computes the posterior expectation for each observation in the graph
}
