% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expontial.covariance.R
\name{posterior.mean.stupid}
\alias{posterior.mean.stupid}
\title{posterior mean calculation not using sparsity}
\usage{
posterior.mean.stupid(theta, graph.obj)
}
\arguments{
\item{theta}{- (sigma_e, sigma, kappa)}

\item{graph.obj}{- graph object}
}
\value{
The posterior mean
}
\description{
posterior mean calculation not using sparsity
}
