% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_inference.R
\name{posterior.mean.stupid}
\alias{posterior.mean.stupid}
\title{posterior mean calculation not using sparsity}
\usage{
posterior.mean.stupid(theta, graph)
}
\arguments{
\item{theta}{\itemize{
\item (sigma_e, sigma, kappa)
}}

\item{graph}{\itemize{
\item metric_graph object
}}
}
\value{
The posterior mean
}
\description{
posterior mean calculation not using sparsity
}
