% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_inference.R
\name{posterior_crossvalidation_covariance}
\alias{posterior_crossvalidation_covariance}
\title{Prediction for models assuming observations at vertices for graph_lme models}
\usage{
posterior_crossvalidation_covariance(object)
}
\arguments{
\item{object}{A fitted model using the \code{graph_lme()} function.}
}
\value{
Vector with all predictions
}
\description{
Prediction for models assuming observations at vertices for graph_lme models
}
\details{
This function does not use sparsity for any model.
}
