% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_inference.R
\name{posterior_crossvalidation_covariance_manual}
\alias{posterior_crossvalidation_covariance_manual}
\title{Prediction for models assuming observations at vertices}
\usage{
posterior_crossvalidation_covariance_manual(
  theta,
  graph,
  data_name,
  model = "alpha1",
  ind = NULL,
  BC = 1
)
}
\arguments{
\item{theta}{Estimated model parameters (sigma_e, tau, kappa)}

\item{graph}{metric graph object}

\item{data_name}{Name of the column of the response variable.}

\item{model}{Type of model: "alpha1" gives SPDE with alpha=1, "GL1" gives
the model based on the graph Laplacian with smoothness 1, "GL2" gives the
model based on the graph Laplacian with smoothness 2, and "isoExp" gives a
model with isotropic exponential covariance}

\item{ind}{Indices for cross validation. It should be a vector of the same
length as the data, with integer values representing each group in the
cross-validation. If NULL, leave-one-out cross validation is performed.}

\item{BC}{which boundary condition to use (0,1) 0 is no adjustment on boundary point
1 is making the boundary condition stationary}
}
\value{
Vector with all predictions
}
\description{
Prediction for models assuming observations at vertices
}
\details{
This function does not use sparsity for any model.
}
