% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_inference.R
\name{posterior_crossvalidation_loo}
\alias{posterior_crossvalidation_loo}
\title{Leave-one-out pseudo-crossvalidation for \code{graph_lme} models assuming observations at
the vertices of metric graphs}
\usage{
posterior_crossvalidation_loo(object, factor = 1, tibble = TRUE)
}
\arguments{
\item{object}{A fitted model using the \code{graph_lme()} function or a named list of fitted objects using the \code{graph_lme()} function.}

\item{factor}{Which factor to multiply the scores. The default is 1.}

\item{tibble}{Return the scores as a \code{tidyr::tibble()}}
}
\value{
Vector with the posterior expectations and variances as well as
mean absolute error (MAE), root mean squared errors (RMSE), and three
negatively oriented proper scoring rules: log-score, CRPS, and scaled
CRPS.
}
\description{
This function performs pseudo-crossvalidation by computing leave-one-out predictions
using the posterior distribution from a fitted model. In pseudo-crossvalidation,
the model parameters are kept fixed at the values estimated from the full dataset
(those provided in the object), rather than re-estimating them for each fold.
}
