% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_lme.R
\name{predict.graph_lme}
\alias{predict.graph_lme}
\title{Prediction of a fractional SPDE using the covariance-based
rational SPDE approximation}
\usage{
\method{predict}{graph_lme}(
  object,
  data,
  repl = NULL,
  compute.variances = FALSE,
  posterior_samples = FALSE,
  n_samples = 100,
  only_latent = FALSE,
  edge_number = "edge_number",
  distance_on_edge = "distance_on_edge",
  normalized = FALSE,
  ...
)
}
\arguments{
\item{object}{The covariance-based rational SPDE approximation,
computed using \code{\link[=matern.operators]{matern.operators()}}}

\item{data}{A \code{data.frame} or a \code{list} containing the covariates, the edge number and the distance on edge
for the locations to obtain the prediction.}

\item{repl}{Which replicates to obtain the prediction. If \code{NULL} predictions will be obtained for all replicates. Default is \code{NULL}.}

\item{compute.variances}{Set to also TRUE to compute the kriging variances.}

\item{posterior_samples}{If \code{TRUE}, posterior samples will be returned.}

\item{n_samples}{Number of samples to be returned. Will only be used if \code{sampling} is \code{TRUE}.}

\item{only_latent}{Should the posterior samples be only given to the laten model?}

\item{edge_number}{Name of the variable that contains the edge number, the default is \code{edge_number}.}

\item{distance_on_edge}{Name of the variable that contains the distance on edge, the default is \code{distance_on_edge}.}

\item{normalized}{Are the distances on edges normalized?}

\item{...}{Not used.}
}
\description{
The function is used for computing kriging predictions based
on data \eqn{Y_i = u(s_i) + \epsilon_i}, where \eqn{\epsilon}{\epsilon}
is mean-zero Gaussian measurement noise and \eqn{u(s)}{u(s)} is defined by
a fractional SPDE \eqn{(\kappa^2 I - \Delta)^{\alpha/2} (\tau u(s)) = W},
where \eqn{W}{W} is Gaussian white noise and \eqn{\alpha = \nu + d/2},
where \eqn{d} is the dimension of the domain.
}
