% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_spde.R
\name{process_rspde_predictions}
\alias{process_rspde_predictions}
\title{Process predictions of \code{rspde_metric_graph} objects obtained by using \code{inlabru}}
\usage{
process_rspde_predictions(pred, graph, PtE = NULL)
}
\arguments{
\item{pred}{The predictions of the field obtained by using \code{inlabru}}

\item{graph}{The original \code{metric_graph} object in which the predictions were obtained.}

\item{PtE}{Normalized locations of the points on the edge.}
}
\value{
A list with predictions.
}
\description{
Auxiliar function to transform the predictions of the field into a plot friendly object.
}
