% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psp_to_metricgraph.R
\name{psp.to.graph}
\alias{psp.to.graph}
\title{Convert a \code{psp} object to a metric graph object}
\usage{
psp.to.graph(psp.object)
}
\arguments{
\item{psp.object}{A \code{psp} object to be converted.}
}
\value{
A metric graph object with edges defined by the segments in the \code{psp} object.
}
\description{
This function converts a \code{psp} object (from the \code{spatstat} package) into a metric graph object.
}
