% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.object.R
\name{Q.exp}
\alias{Q.exp}
\title{Building the precision matrix for the expontial case on each vertex}
\usage{
Q.exp(theta, V, EtV, El)
}
\arguments{
\item{theta}{- kappa}

\item{V}{vertex position}

\item{EtV}{[,2-3] index of upper and lower edge}

\item{El}{length of each vertex}
}
\value{
Q (precision matrix)
}
\description{
Building the precision matrix for the expontial case on each vertex
}
