% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spde_precision.R
\name{r_1}
\alias{r_1}
\title{The exponential covariance}
\usage{
r_1(D, theta)
}
\arguments{
\item{D}{vector or matrix with distances}

\item{theta}{parameters sigma,kappa}
}
\description{
The exponential covariance
}
