% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance_alpha1.R
\name{r_1}
\alias{r_1}
\title{The exponential covariance}
\usage{
r_1(D, kappa, sigma)
}
\arguments{
\item{D}{vector or matrix with distances}

\item{kappa}{parameter kappa}

\item{sigma}{parameter sigma}
}
\description{
The exponential covariance
}
