% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expontial.covariance.R
\name{r_1_circle}
\alias{r_1_circle}
\title{the expontial circular covariance}
\usage{
r_1_circle(t, l_e, theta)
}
\arguments{
\item{t}{- locations}

\item{l_e}{- length circle}

\item{theta}{- kappa, sigma}
}
\description{
the expontial circular covariance
}
