% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance_alpha1.R
\name{r_1_circle}
\alias{r_1_circle}
\title{the Whittle--Matern covariance with alpha=1 on a circle}
\usage{
r_1_circle(t, l_e, kappa, sigma)
}
\arguments{
\item{t}{locations}

\item{l_e}{circle perimeter}

\item{kappa}{parameter kappa}

\item{sigma}{parameter sigma}
}
\description{
the Whittle--Matern covariance with alpha=1 on a circle
}
