% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance_alpha2.R
\name{r_2}
\alias{r_2}
\title{The Matern covariance with alpha = 2}
\usage{
r_2(D, kappa, sigma, deriv = 0)
}
\arguments{
\item{D}{vector or matrix with distances}

\item{kappa}{parameter kappa}

\item{sigma}{parameter sigma}

\item{deriv}{(0,1,2) no derivative, first, or second order}
}
\description{
The Matern covariance with alpha = 2
}
