% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spde_precision.R
\name{r_2}
\alias{r_2}
\title{The Matern covariance with alpha=2}
\usage{
r_2(D, theta, deriv = 0)
}
\arguments{
\item{D}{vector or matrix with distances}

\item{theta}{(sigma, kappa)}

\item{deriv}{(0,1,2) no derivative, first, or second order}
}
\description{
The Matern covariance with alpha=2
}
