% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{select.metric_graph_data}
\alias{select.metric_graph_data}
\alias{select}
\title{A version of \code{dplyr::select()} function for datasets on metric graphs}
\usage{
\method{select}{metric_graph_data}(.data, ...)
}
\arguments{
\item{.data}{The data list or \code{tidyr::tibble} obtained from a metric graph object.}

\item{...}{Additional parameters to be passed to \code{dplyr::select()}.}
}
\value{
A \code{tidyr::tibble} with the resulting selected columns.
}
\description{
Selects columns on metric graphs, while keeps the spatial positions.
}
