% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/space_time.R
\name{simulate.spacetime}
\alias{simulate.spacetime}
\title{space-time simulation based on implicit Euler discretization in time}
\usage{
\method{simulate}{spacetime}(graph, t, kappa, rho, gamma, alpha, beta, sigma, u0, BC = 0)
}
\arguments{
\item{graph}{metric_graph object}

\item{t}{vector of time points}

\item{kappa}{range parameter kappa}

\item{rho}{drift parameter}

\item{gamma}{temporal range parameter}

\item{alpha}{smoothness parameter (integer) for spatial operator}

\item{beta}{smoothness parameter (integer) for Q-Wiener process}

\item{sigma}{variance parameter}

\item{u0}{starting value}
}
\value{
Precision matrix
}
\description{
Simulation with starting value u0
}
