% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psp_to_metricgraph.R
\name{stlpp.to.graph}
\alias{stlpp.to.graph}
\title{function to export a stlpp object, from stlnpp, to metric graph object}
\usage{
stlpp.to.graph(stlpp.obj)
}
\description{
function to export a stlpp object, from stlnpp, to metric graph object
}
