% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.object.R
\name{vertex.to.line}
\alias{vertex.to.line}
\title{Assume that there is only one line each Lines}
\usage{
vertex.to.line(Lines)
}
\arguments{
\item{Lines}{object}
}
\value{
A list with elements V, EtV, El
}
\description{
Assume that there is only one line each Lines
}
